+++ 01_Backend/backend/src/main/java/org/example/backend/repository/order/OrderCancelRepository.java
@@ -1,13 +1,11 @@
-package org.example.backend.repository;
+package org.example.backend.repository.order;
 
-import org.example.backend.model.entity.OrderDetail;
+import org.example.backend.model.entity.OrderCancel;
 import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.List;
 
 /**
  * packageName : org.example.backend.repository
- * fileName : OrderDetailRepository
+ * fileName : OrderCancelRepository
  * author : kimtaewan
  * date : 2024-04-25
  * description :
@@ -18,6 +16,5 @@ import java.util.List;
  * -----------------------------------------------------------
  * 2024-04-25         kimtaewan          최초 생성
  */
-public interface OrderDetailRepository extends JpaRepository<OrderDetail, Long> {
-    List<OrderDetail> findByOrderId(Long orderId);
+public interface OrderCancelRepository extends JpaRepository<OrderCancel, Long> {
 }
