+++ 01_Backend/backend/src/main/java/org/example/backend/repository/order/OrderDetailRepository.java
@@ -1,11 +1,13 @@
-package org.example.backend.repository;
+package org.example.backend.repository.order;
 
-import org.example.backend.model.entity.ProductImage;
+import org.example.backend.model.entity.OrderDetail;
 import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.List;
 
 /**
  * packageName : org.example.backend.repository
- * fileName : ProductImageRepository
+ * fileName : OrderDetailRepository
  * author : kimtaewan
  * date : 2024-04-25
  * description :
@@ -16,5 +18,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
  * -----------------------------------------------------------
  * 2024-04-25         kimtaewan          최초 생성
  */
-public interface ProductImageRepository extends JpaRepository<ProductImage, Long> {
+public interface OrderDetailRepository extends JpaRepository<OrderDetail, Long> {
+    List<OrderDetail> findByOrderId(Long orderId);
 }
